Move some code to BitmapRenderer.Draw and SvgRenderer.Draw
